kernel/svc: Split implementations into separate files
